Fix mem. seg. fault error in jsoncpp lib
The root cause of the issue is a stack overflow which is occurring
due to recursive call of deeply nested areas and objects. We have
readValue(..)  which calls readObject(..) or readArray(..),
which call readValue(...). The starting point is a call to parse(..).
In order to fix the issue a count var. is implemented, which
measure how deep is our recursion. When recursion threshold is
exceeded we return invalid messege.